+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/txClient/TerminologyClientR3.java
@@ -127,6 +127,11 @@ public class TerminologyClientR3 implements ITerminologyClient {
     client.setTimeoutFactor(i);
     return this;
   }
+  
+  @Override
+  public ToolingClientLogger getLogger() {
+    return client.getLogger();
+  }
 
   @Override
   public ITerminologyClient setLogger(ToolingClientLogger txLog) {
@@ -228,6 +233,13 @@ public class TerminologyClientR3 implements ITerminologyClient {
   @Override
   public int getUseCount() {
     return client.getUseCount();
+  }
+
+  @Override
+  public Bundle search(String type, String criteria) {   
+    org.hl7.fhir.dstu3.model.Bundle result = client.search(type, criteria);
+    return result == null ? null : (Bundle) VersionConvertorFactory_30_50.convertResource(result);
+
   }
 
 }