+++ org.hl7.fhir.convertors/src/main/java/org/hl7/fhir/convertors/txClient/TerminologyClientR4.java
@@ -139,6 +139,11 @@ public class TerminologyClientR4 implements ITerminologyClient {
   }
 
   @Override
+  public ToolingClientLogger getLogger() {
+    return client.getLogger();
+  }
+
+  @Override
   public ITerminologyClient setLogger(ToolingClientLogger txLog) {
     client.setLogger(txLog);
     return this;
@@ -240,6 +245,12 @@ public class TerminologyClientR4 implements ITerminologyClient {
   @Override
   public int getUseCount() {
     return client.getUseCount();
+  }
+
+  @Override
+  public Bundle search(String type, String criteria) {    
+    org.hl7.fhir.r4.model.Bundle result = client.search(type, criteria);
+    return result == null ? null : (Bundle) VersionConvertorFactory_40_50.convertResource(result);
   }