+++ org.hl7.fhir.dstu3/src/main/java/org/hl7/fhir/dstu3/utils/client/FHIRToolingClient.java
@@ -654,5 +654,10 @@ public class FHIRToolingClient extends FHIRBaseToolingClient {
     useCount++;    
   }
 
+  public Bundle search(String type, String criteria) {
+    recordUse();
+    return fetchFeed(Utilities.pathURL(base, type+criteria));
+  }
+  
 }