+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/conformance/profile/BindingResolution.java
@@ -10,4 +10,6 @@ public class BindingResolution {
   }
   public String display;
   public String url;
+  public String uri;
+  public boolean external;
 }