+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/context/SimpleWorkerContext.java
@@ -41,6 +41,7 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.Set;
+import java.util.UUID;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
 
@@ -51,6 +52,7 @@ import org.apache.commons.io.IOUtils;
 import org.hl7.fhir.exceptions.DefinitionException;
 import org.hl7.fhir.exceptions.FHIRException;
 import org.hl7.fhir.exceptions.FHIRFormatError;
+import org.hl7.fhir.exceptions.TerminologyServiceException;
 import org.hl7.fhir.r5.context.CanonicalResourceManager.CanonicalResourceProxy;
 import org.hl7.fhir.r5.context.ILoggingService.LogCategory;
 import org.hl7.fhir.r5.formats.IParser;
@@ -64,6 +66,7 @@ import org.hl7.fhir.r5.model.StructureMap.StructureMapModelMode;
 import org.hl7.fhir.r5.model.StructureMap.StructureMapStructureComponent;
 import org.hl7.fhir.r5.terminologies.JurisdictionUtilities;
 import org.hl7.fhir.r5.terminologies.client.ITerminologyClient;
+import org.hl7.fhir.r5.terminologies.client.TerminologyClientContext;
 import org.hl7.fhir.r5.terminologies.client.TerminologyClientManager;
 import org.hl7.fhir.r5.terminologies.client.TerminologyClientManager.ITerminologyClientFactory;
 import org.hl7.fhir.r5.terminologies.client.TerminologyClientR5;
@@ -237,7 +240,7 @@ public class SimpleWorkerContext extends BaseWorkerContext implements IWorkerCon
     }
 
     private SimpleWorkerContext build(SimpleWorkerContext context) throws IOException {
-      context.initTS(terminologyCachePath);
+      context.initTxCache(terminologyCachePath);
       context.setUserAgent(userAgent);
       context.setLogger(loggingService);
       context.cacheResource(new org.hl7.fhir.r5.formats.JsonParser().parse(MagicResources.spdxCodesAsData()));
@@ -247,7 +250,7 @@ public class SimpleWorkerContext extends BaseWorkerContext implements IWorkerCon
     public SimpleWorkerContext fromPackage(NpmPackage pi) throws IOException, FHIRException {
       SimpleWorkerContext context = getSimpleWorkerContextInstance();
       context.setAllowLoadingDuplicates(allowLoadingDuplicates);
-      context.terminologyClientManager = new TerminologyClientManager(TerminologyClientR5.factory());
+      context.terminologyClientManager.setFactory(TerminologyClientR5.factory());
       context.loadFromPackage(pi, null);
       return build(context);
     }
@@ -328,30 +331,48 @@ public class SimpleWorkerContext extends BaseWorkerContext implements IWorkerCon
       loadBytes(name, stream);
   }
 
-  public String connectToTSServer(ITerminologyClientFactory factory, ITerminologyClient client, String log) {
+  public void connectToTSServer(ITerminologyClientFactory factory, ITerminologyClient client) {
+    terminologyClientManager.setFactory(factory);
+    if (txLog == null) {
+      txLog = client.getLogger();
+    }
+    TerminologyClientContext tcc = terminologyClientManager.setMasterClient(client);
+    txLog("Connect to "+client.getAddress());
     try {
-      txLog("Connect to "+client.getAddress());
+      tcc.initialize();  
+    } catch (Exception e) {
+      if (canRunWithoutTerminology) {
+        noTerminologyServer = true;
+        logger.logMessage("==============!! Running without terminology server !! ==============");
+        if (terminologyClientManager.getMasterClient() != null) {
+          logger.logMessage("txServer = "+ terminologyClientManager.getMasterClient().getId());
+          logger.logMessage("Error = "+e.getMessage()+"");
+        }
+        logger.logMessage("=====================================================================");
+      } else {
+        e.printStackTrace();
+        throw new TerminologyServiceException(e);
+      }
+    }      
+  }
+  
+  public void connectToTSServer(ITerminologyClientFactory factory, String address, String software, String log) {
+    try {
       terminologyClientManager.setFactory(factory);
-      terminologyClientManager.setMasterClient(client);
       if (log != null && (log.endsWith(".htm") || log.endsWith(".html"))) {
         txLog = new HTMLClientLogger(log);
       } else {
         txLog = new TextClientLogger(log);
-      }
-      terminologyClientManager.setLogger(txLog);
-      terminologyClientManager.setUserAgent(userAgent);
-
-      final CapabilityStatement capabilitiesStatementQuick = txCache.hasCapabilityStatement() ? txCache.getCapabilityStatement() : terminologyClientManager.getMasterClient().getCapabilitiesStatementQuick();
-      txCache.cacheCapabilityStatement(capabilitiesStatementQuick);
-
-      final TerminologyCapabilities capabilityStatement = txCache.hasTerminologyCapabilities() ? txCache.getTerminologyCapabilities() : terminologyClientManager.getMasterClient().getTerminologyCapabilities();
-      txCache.cacheTerminologyCapabilities(capabilityStatement);
-
-      setTxCaps(capabilityStatement);
-      return capabilitiesStatementQuick.getSoftware().getVersion();
+      }      
+      ITerminologyClient client = factory.makeClient("tx-server", address, software, txLog);
+      // txFactory.makeClient("Tx-Server", txServer, "fhir/publisher", null)
+//      terminologyClientManager.setLogger(txLog);
+//      terminologyClientManager.setUserAgent(userAgent);
+      connectToTSServer(factory, client);
+      
     } catch (Exception e) {
       e.printStackTrace();
-      throw new FHIRException(formatMessage(canNoTS ? I18nConstants.UNABLE_TO_CONNECT_TO_TERMINOLOGY_SERVER_USE_PARAMETER_TX_NA_TUN_RUN_WITHOUT_USING_TERMINOLOGY_SERVICES_TO_VALIDATE_LOINC_SNOMED_ICDX_ETC_ERROR__ : I18nConstants.UNABLE_TO_CONNECT_TO_TERMINOLOGY_SERVER, e.getMessage(), client.getAddress()), e);
+      throw new FHIRException(formatMessage(canNoTS ? I18nConstants.UNABLE_TO_CONNECT_TO_TERMINOLOGY_SERVER_USE_PARAMETER_TX_NA_TUN_RUN_WITHOUT_USING_TERMINOLOGY_SERVICES_TO_VALIDATE_LOINC_SNOMED_ICDX_ETC_ERROR__ : I18nConstants.UNABLE_TO_CONNECT_TO_TERMINOLOGY_SERVER, e.getMessage(), address), e);
     }
   }