+++ org.hl7.fhir.r5/src/main/java/org/hl7/fhir/r5/test/utils/TestingUtilities.java
@@ -140,7 +140,7 @@ public class TestingUtilities extends BaseTestingUtilities {
     if (!Utilities.noString(s))
       return s;
     s = "C:\\work\\org.hl7.fhir\\build";
-    // FIXME: change this back
+    // #TODO - what should we do with this?
     s = "/Users/jamesagnew/git/fhir";
     if (new File(s).exists())
       return s;