+++ RELEASE_NOTES.md
@@ -1,7 +1,12 @@
 ## Validator Changes
 
 * Fix issues with multiple terminology servers (should be live soon)
+* Rework terminology system & cache to support new terminology server ecosystem
+* Handle value set with exclusion filters when expanding
+* Fix to consistent error messages when value sets can't be resolved
 
 ## Other code changes
 
-* no changes
+* Fix bug processing empty structure maps
+* Performance improvements to common IO routines, with test cases - see https://blog.qligier.ch/posts/2024-improving-perf-ig-publisher-io/ (thanks to Quentin Ligier)
+